requirements.txt
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
supabase==2.0.3
rich==13.7.0
requests==2.31.0
